requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==4.2.19
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.1.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
openai==0.28.0
packaging==24.2
propcache==0.3.0
PyJWT==2.9.0
PyMySQL==1.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
